pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robot-visual-memory"
version = "0.1.0"
description = "Real-time visual place recognition and scene memory for robots using CLIP, Qdrant, and ROS2"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "open-clip-torch>=2.20.0",
    "qdrant-client>=1.7.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0"]
ros2 = ["rclpy"]

[project.scripts]
robot-visual-memory = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["src*"]
